Move minmax time function to service
